% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventure.R
\docType{package}
\name{eventure-package}
\alias{eventure}
\alias{eventure-package}
\title{Describe events}
\description{
\code{eventure} provides a dsl for describing and handling events.
}
\author{
\strong{Maintainer}: Nathan Teetor \email{nathanteetor@gmail.com}

}
